@@ -1,41 +1,68 @@
 [build-system]
-requires = ["setuptools>=42", "wheel", "build>=0.7.0"]
+requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-# This section is defined by PEP 621
-name = "macos-versiontracker"
-dynamic = ["version"]
-description = "CLI versiontracker and update tool for macOS"
+name = "homebrew-versiontracker"
+version = "0.1.0-alpha.1"
+description = "A command-line tool for tracking and managing applications installed outside of the Mac App Store"
 readme = "README.md"
-requires-python = ">=3.9"
-license = "MIT"
-authors = [{ name = "docdyhr", email = "thomas@dyhr.com" }]
-keywords = ["versiontracker", "macos", "update", "homebrew", "mac-apps"]
+license = {text = "MIT"}
+requires-python = ">=3.12"
+authors = [
+    {name = "docdyhr", email = "thomas@dyhr.com"}
+]
+keywords = ["macos", "homebrew", "version", "tracking", "updates", "cask"]
 classifiers = [
-    "Development Status :: 4 - Beta",
+    "Development Status :: 3 - Alpha",
     "Environment :: Console",
-    "Intended Audience :: Developers",
     "Intended Audience :: System Administrators",
+    "Intended Audience :: End Users/Desktop",
+    "License :: OSI Approved :: MIT License",
     "Operating System :: MacOS :: MacOS X",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
-    "Programming Language :: Python :: 3.13",
+    "Topic :: System :: Software Distribution",
     "Topic :: System :: Systems Administration",
     "Topic :: Utilities",
 ]
+
 dependencies = [
-    "fuzzywuzzy>=0.18.0",
-    "rapidfuzz>=3.0.0",
-    "tqdm>=4.66.0",
-    "PyYAML>=6.0",
-    "termcolor>=2.3.0",
+    "pyyaml>=6.0",
+    "click>=8.0",
+    "tqdm>=4.65",
+    "psutil>=5.9",
     "tabulate>=0.9.0",
-    "psutil>=6.1.0,<8.0.0",
     "aiohttp>=3.8.0",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.4",
+    "pytest-cov>=4.1",
+    "pytest-timeout>=2.1",
+    "pytest-mock>=3.11",
+    "black>=23.7",
+    "ruff>=0.1",
+    "mypy>=1.5",
+    "types-PyYAML>=6.0",
+    "build>=0.10",
+    "wheel>=0.41",
+    "twine>=4.0",
+]
+test = [
+    "pytest>=7.4",
+    "pytest-cov>=4.1",
+    "pytest-timeout>=2.1",
+    "pytest-mock>=3.11",
+]
+security = [
+    "bandit[toml]>=1.7",
+    "pip-audit>=2.6",
+    "safety>=2.3",
+]
+fuzzy = [
+    "rapidfuzz>=3.0",
 ]
 
 [project.urls]
@@ -44,36 +71,26 @@ Homepage = "https://github.com/docdyhr/versiontracker"
 Documentation = "https://github.com/docdyhr/versiontracker"
 "Source Code" = "https://github.com/docdyhr/versiontracker"
 Changelog = "https://github.com/docdyhr/versiontracker/blob/main/CHANGELOG.md"
-PyPI = "https://pypi.org/project/macos-versiontracker/"
+PyPI = "https://pypi.org/project/homebrew-versiontracker/"
 
 [project.scripts]
-versiontracker = "versiontracker.__main__:main"
-macos-versiontracker = "versiontracker.__main__:main"
+versiontracker = "versiontracker.cli:main"
 
-[tool.setuptools.packages.find]
-where = ["."]
-include = ["versiontracker*"]
-exclude = [
-    "images*",
-    "tmp_test*",
-    "tests*",
-    "docs*",
-    "htmlcov*",
-    "tools*",
-    "handlers*",
-    "node_modules*",
-]
+[tool.setuptools]
+packages = ["versiontracker", "versiontracker.handlers"]
+
+[tool.setuptools.package-data]
+versiontracker = ["py.typed"]
+
+[tool.black]
+line-length = 88
+target-version = ["py312"]
+include = '\.pyi?$'
 
 [tool.ruff]
-# Define the target Python version.
-# We support Python 3.13 and later.
-target-version = "py39"
+line-length = 120  # AI-friendly length as per CLAUDE.md guidelines
+target-version = "py312"  # Updated to match project requirements
 
-# Line length configuration - harmonized for AI assistant best practices
-# 120 characters provides good readability while being AI-friendly
-line-length = 120
-
-# Exclude the venv directory.
 exclude = [
     ".bzr",
     ".direnv",
