@@ -1,1 +1,9 @@
-"""Test package for VersionTracker."""
+"""Test suite for VersionTracker."""
+
+import sys
+from pathlib import Path
+
+# Add the project root to the Python path
+project_root = Path(__file__).parent.parent
+if str(project_root) not in sys.path:
+    sys.path.insert(0, str(project_root))
