@@ -1,66 +1,64 @@
-"""Tests for the config module."""
+"""
+Test module for configuration functionality.
 
+This module provides basic tests for the config module.
+"""
+
+import logging
 import os
+import tempfile
 import unittest
 from unittest.mock import patch
 
-from versiontracker.config import Config
+from versiontracker.config import Config, get_config
 
 
-class TestConfig(unittest.TestCase):
-    """Test cases for the Config class."""
+class TestConfig:
+    """Tests for configuration management."""
 
-    def setUp(self):
-        """Set up test fixtures, if any."""
-        # Store original environment variables and clear test-related ones
-        self.original_env = {}
-        env_vars_to_clear = [var for var in os.environ.keys() if var.startswith("VERSIONTRACKER_")]
-        for var in env_vars_to_clear:
-            self.original_env[var] = os.environ[var]
-            del os.environ[var]
-
-    def tearDown(self):
-        """Clean up test fixtures."""
-        # Restore original environment variables
-        for var, value in self.original_env.items():
-            os.environ[var] = value
-
-    def test_init_default_values(self):
-        """Test that default values are set correctly."""
-        # Create a Config with a non-existent file to get defaults only
-        config = Config(config_file="/nonexistent/config.yaml")
-        self.assertEqual(config.get("api_rate_limit"), 3)
-        self.assertEqual(len(config.get_blacklist()), 8)
-        self.assertTrue(config.get("show_progress"))
-
-    def test_set_get(self):
-        """Test setting and getting values."""
+    def test_default_config(self):
+        """Test that default configuration loads properly."""
         config = Config()
-        config.set("test_key", "test_value")
-        self.assertEqual(config.get("test_key"), "test_value")
-        self.assertIsNone(config.get("nonexistent_key"))
-        self.assertEqual(config.get("nonexistent_key", "default"), "default")
+        assert config is not None
+        # Test some expected defaults
+        assert hasattr(config, "max_workers")
+        assert config.max_workers > 0
 
-    def test_is_blacklisted(self):
-        """Test blacklist functionality."""
-        config = Config()
-        config.set("blacklist", ["App1", "App2"])
-        self.assertTrue(config.is_blacklisted("App1"))
-        self.assertTrue(config.is_blacklisted("app1"))  # Case insensitive
-        self.assertFalse(config.is_blacklisted("App3"))
+    def test_get_config_singleton(self):
+        """Test that get_config returns the same instance."""
+        config1 = get_config()
+        config2 = get_config()
+        assert config1 is config2
 
-    @patch.dict(os.environ, {"VERSIONTRACKER_API_RATE_LIMIT": "5"})
-    def test_env_api_rate_limit(self):
-        """Test loading API rate limit from environment."""
-        config = Config()
-        self.assertEqual(config.get("api_rate_limit"), 5)
+    def test_config_from_file(self):
+        """Test loading configuration from a YAML file."""
+        with tempfile.NamedTemporaryFile(mode="w", suffix=".yml", delete=False) as f:
+            f.write("max_workers: 10\n")
+            f.write("cache_dir: /tmp/test_cache\n")
+            temp_file = f.name
 
-    @patch.dict(os.environ, {"VERSIONTRACKER_DEBUG": "true"})
-    def test_env_debug(self):
-        """Test loading debug mode from environment."""
-        import logging
+        try:
+            config = Config(config_file=temp_file)
+            assert config.max_workers == 10
+            assert config.cache_dir == "/tmp/test_cache"
+        finally:
+            os.unlink(temp_file)
 
-        config = Config()
+    def test_config_from_env(self):
+        """Test that environment variables override config file."""
+        # Save original env
+        original_workers = os.environ.get("VERSIONTRACKER_MAX_WORKERS")
+
+        try:
+            os.environ["VERSIONTRACKER_MAX_WORKERS"] = "5"
+            config = Config()
+            assert config.max_workers == 5
+        finally:
+            # Restore original env
+            if original_workers is not None:
+                os.environ["VERSIONTRACKER_MAX_WORKERS"] = original_workers
+            else:
+                os.environ.pop("VERSIONTRACKER_MAX_WORKERS", None)
         self.assertEqual(config.get("log_level"), logging.DEBUG)
 
     @patch.dict(os.environ, {"VERSIONTRACKER_BLACKLIST": "App1,App2,App3"})
