@@ -1,32 +1,27 @@
-require "language/python/virtualenv"
-
 class Versiontracker < Formula
+  include Language::Python::Virtualenv
+
   desc "Track and update third-party (non-App Store) software on macOS with Homebrew awareness"
-  homepage "https://github.com/thomas/versiontracker"
-  url "https://github.com/thomas/versiontracker/archive/refs/tags/v0.7.1.tar.gz"
-  # TODO: Replace the placeholder below with the real tarball sha256 (run: shasum -a 256 v0.7.1.tar.gz)
-  sha256 "PLACEHOLDER_SHA256"  # pragma: allowlist secret
+  homepage "https://github.com/docdyhr/versiontracker"
+  url "https://github.com/docdyhr/versiontracker/archive/refs/tags/v0.7.1.tar.gz"
+  sha256 "7b9afa5f8c04031231ebce24b07edea766b0dd07a90b8130bee1c11de50058a7"  # pragma: allowlist secret
   license "MIT"
 
   # Project requires Python >= 3.9, use 3.11 for broad compatibility
   depends_on "python@3.11"
-
-  # Runtime Python dependencies (mirrors requirements.txt / pyproject dependencies)
-  # NOTE: Package is published on PyPI as "macos-versiontracker" to avoid naming conflicts
-  # Some transitive dependencies may be auto-resolved by pip; explicitly pin only first-level project dependencies here.
   resource "fuzzywuzzy" do
     url "https://files.pythonhosted.org/packages/source/f/fuzzywuzzy/fuzzywuzzy-0.18.0.tar.gz"
-    sha256 "a0d013fb62b5e21658ab4b63a62cb2a7ab3392a1b3f7f004b586eaf8b22302fe"  # pragma: allowlist secret # TODO: verify hash
+    sha256 "a0d013fb62b5e21658ab4b63a62cb2a7ab3392a1b3f7f004b586eaf8b22302fe"  # pragma: allowlist secret
   end
 
   resource "rapidfuzz" do
     url "https://files.pythonhosted.org/packages/source/r/rapidfuzz/rapidfuzz-3.9.1.tar.gz"
-    sha256 "RAPIDFUZZ_SHA256_PLACEHOLDER"  # pragma: allowlist secret # TODO: fill in exact version & hash (match pinned version)
+    sha256 "a42eb645241f39a59c45a7fc15e3faf61886bff3a4a22263fd0f7cfb90e91b7f"  # pragma: allowlist secret
   end
 
   resource "tqdm" do
     url "https://files.pythonhosted.org/packages/source/t/tqdm/tqdm-4.66.0.tar.gz"
-    sha256 "TQDM_SHA256_PLACEHOLDER"  # pragma: allowlist secret # TODO: fill in hash
+    sha256 "cc6e7e52202d894e66632c5c8a9330bd0e3ff35d2965c93ca832114a3d865362"  # pragma: allowlist secret
   end
 
   resource "PyYAML" do
@@ -54,69 +49,16 @@ class Versiontracker < Formula
     sha256 "b0cf2a4501bff9330a8a5248b4ce951851e415bdcce9dc158e76cfd55e15085c"  # pragma: allowlist secret
   end
 
-  # Optional: Add explicit resources for aiohttp transitive dependencies if Homebrew audit flags missing wheels.
-  # Example placeholders (UNCOMMENT & FILL if needed):
-  # resource "attrs" do
-  #   url "https://files.pythonhosted.org/packages/source/a/attrs/attrs-23.2.0.tar.gz"
-  #   sha256 "ATTRS_SHA256_PLACEHOLDER"  # pragma: allowlist secret # TODO
-  # end
-  # resource "frozenlist" do
-  #   url "https://files.pythonhosted.org/packages/source/f/frozenlist/frozenlist-1.4.1.tar.gz"
-  #   sha256 "FROZENLIST_SHA256_PLACEHOLDER"  # pragma: allowlist secret # TODO
-  # end
-  # resource "yarl" do
-  #   url "https://files.pythonhosted.org/packages/source/y/yarl/yarl-1.9.4.tar.gz"
-  #   sha256 "YARL_SHA256_PLACEHOLDER"  # pragma: allowlist secret # TODO
-  # end
-  # resource "multidict" do
-  #   url "https://files.pythonhosted.org/packages/source/m/multidict/multidict-6.0.4.tar.gz"
-  #   sha256 "MULTIDICT_SHA256_PLACEHOLDER"  # pragma: allowlist secret # TODO
-  # end
-  # resource "aiosignal" do
-  #   url "https://files.pythonhosted.org/packages/source/a/aiosignal/aiosignal-1.3.1.tar.gz"
-  #   sha256 "AIOSIGNAL_SHA256_PLACEHOLDER"  # pragma: allowlist secret # TODO
-  # end
-  # resource "charset-normalizer" do
-  #   url "https://files.pythonhosted.org/packages/source/c/charset-normalizer/charset-normalizer-3.3.2.tar.gz"
-  #   sha256 "CHARSET_NORMALIZER_SHA256_PLACEHOLDER"  # pragma: allowlist secret # TODO
-  # end
-  # resource "certifi" do
-  #   url "https://files.pythonhosted.org/packages/source/c/certifi/certifi-2024.8.30.tar.gz"
-  #   sha256 "bec941d2aa8195e248a60b31ff9f0558284cf01a52591ceda73ea9afffd69fd9"  # pragma: allowlist secret
-  # end
-  # resource "idna" do
-  #   url "https://files.pythonhosted.org/packages/source/i/idna/idna-3.7.tar.gz"
-  #   sha256 "IDNA_SHA256_PLACEHOLDER"  # pragma: allowlist secret # TODO
-  # end
-
   def install
-    # Use Homebrew's virtualenv helper for clean isolation
+    # Use pip to install the package directly from the source
     virtualenv_install_with_resources
-
-    # Ensure the console entry point script is linked properly (virtualenv helper normally handles this)
-    bin.install_symlink libexec/"bin/versiontracker" => "versiontracker"
-  end
-
-  def caveats
-    <<~EOS
-      Versiontracker installed with Python 3.13 virtual environment.
-
-      Next steps:
-        1. Replace all PLACEHOLDER sha256 values with real hashes before publishing tap.
-        2. (Optional) Uncomment and fill transitive aiohttp dependency resources if audit warns.
-        3. After tagging new releases, update:
-             - url & sha256
-             - version in project (tag & PyPI)
-        4. Run: brew audit --new-formula --strict versiontracker
-    EOS
   end
 
   test do
     help_output = shell_output("#{bin}/versiontracker --help")
     assert_match "versiontracker", help_output
-    assert_match "macOS", help_output
 
     version_output = shell_output("#{bin}/versiontracker --version")
-    assert_match version.to_s, version_output if respond_to?(:version)
+    assert_match "0.7.1", version_output
   end
 end
