@@ -1,311 +1,287 @@
-"""Command Line Interface for VersionTracker."""
+"""
+Command-line interface for VersionTracker.
+
+This module provides the main entry point for the versiontracker command.
+"""
 
 import argparse
 import sys
-import textwrap
-from datetime import datetime
+import traceback
+from typing import Any
+
+import click
 
 from versiontracker import __version__
-from versiontracker.deprecation import warn_deprecated_flag
+from versiontracker.exceptions import VersionTrackerError
 
 
-def get_arguments():
-    """Return a dict of command line arguments (cli).
-
-    Returns:
-        argparse.Namespace: The parsed command line arguments
+@click.command()
+@click.version_option(version=__version__, prog_name="versiontracker")
+@click.option(
+    "--config",
+    "-c",
+    type=click.Path(exists=True),
+    help="Path to configuration file",
+)
+@click.option(
+    "--verbose",
+    "-v",
+    is_flag=True,
+    help="Enable verbose output",
+)
+def main(config: str | None, verbose: bool) -> None:
     """
+    VersionTracker - Application version management for macOS.
+
+    A command-line tool for tracking and managing applications installed
+    outside of the Mac App Store, with Homebrew cask integration.
+    """
+    try:
+        # Load configuration
+
+        from versiontracker.config import Config, get_config
+
+        if config:
+            cfg = Config(config_file=config)
+        else:
+            cfg = get_config()
+
+        if verbose:
+            click.echo(f"VersionTracker v{__version__}")
+            click.echo(f"Configuration loaded from: {config or 'defaults'}")
+            max_workers = getattr(cfg, "max_workers", "N/A")
+            cache_dir = getattr(cfg, "cache_dir", "N/A")
+            click.echo(f"Max workers: {max_workers}")
+            click.echo(f"Cache directory: {cache_dir}")
+
+        # Placeholder for main functionality
+        click.echo(f"VersionTracker v{__version__} - Ready for commands")
+        click.echo("Use --help for available options")
+
+    except VersionTrackerError as e:
+        click.echo(f"Error: {e}", err=True)
+        sys.exit(1)
+    except Exception as e:
+        # Intentional broad exception catch to ensure all errors
+        # are reported to the user
+        click.echo(f"Unexpected error: {e}", err=True)
+        if verbose:
+            traceback.print_exc()
+        sys.exit(1)
+
+
+def get_arguments() -> Any:
+    """Parse command line arguments."""
     parser = argparse.ArgumentParser(
+        description=("VersionTracker - Application version management for macOS"),
         formatter_class=argparse.RawDescriptionHelpFormatter,
-        # Copyright year automatically updates to current year
-        epilog=textwrap.dedent(
-            f"""\
-         thomas@dyhr.com {datetime.now().year}
-         """
-        ),
     )
 
-    # Debug mode
-    parser.add_argument("-D", "--debug", dest="debug", action="store_true", help="turn on DEBUG mode")
-
-    # Performance options
-    performance_group = parser.add_argument_group("Performance options")
-    performance_group.add_argument(
-        "--rate-limit",
-        dest="rate_limit",
-        type=int,
-        help="API rate limit in seconds (default: 3)",
-    )
-    performance_group.add_argument(
-        "--max-workers",
-        dest="max_workers",
-        type=int,
-        help="Maximum number of worker threads (default: 10)",
-    )
-    performance_group.add_argument(
-        "--no-progress",
-        dest="no_progress",
+    # Main action flags (mutually exclusive for actions)
+    action_group = parser.add_mutually_exclusive_group()
+    action_group.add_argument(
+        "--apps",
         action="store_true",
-        help="Disable progress bars",
+        help="List installed applications",
     )
-    performance_group.add_argument(
-        "--profile",
-        dest="profile",
+    action_group.add_argument(
+        "--brews",
         action="store_true",
-        help="Enable performance profiling and report",
+        help="List Homebrew casks",
     )
-    performance_group.add_argument(
-        "--detailed-profile",
-        dest="detailed_profile",
+    action_group.add_argument(
+        "--recom",
+        "--recommend",
+        dest="recom",
         action="store_true",
-        help="Include detailed profiling information in report (requires --profile)",
+        help="Recommend Homebrew casks for installed apps",
+    )
+    action_group.add_argument(
+        "--strict-recom",
+        "--strict-recommend",
+        dest="strict_recom",
+        action="store_true",
+        help="Strict recommendations for Homebrew casks",
+    )
+    action_group.add_argument(
+        "--check-outdated",
+        dest="check_outdated",
+        action="store_true",
+        help="Check for outdated applications",
     )
 
-    # UI options
-    ui_group = parser.add_argument_group("UI options")
-    ui_group.add_argument(
-        "--no-color",
-        dest="no_color",
+    # Auto-update management (mutually exclusive)
+    auto_update_group = parser.add_mutually_exclusive_group()
+    auto_update_group.add_argument(
+        "--blacklist-auto-updates",
+        dest="blacklist_auto_updates",
         action="store_true",
-        help="Disable colored output",
+        help="Add all applications with auto-updates to the blocklist",
     )
-    ui_group.add_argument(
-        "--no-resource-monitor",
-        dest="no_resource_monitor",
+    auto_update_group.add_argument(
+        "--blocklist-auto-updates",
+        dest="blocklist_auto_updates",
         action="store_true",
-        help="Disable system resource monitoring",
+        help="Blocklist applications with auto-update features",
     )
-    ui_group.add_argument(
-        "--no-adaptive-rate",
-        dest="no_adaptive_rate",
+    auto_update_group.add_argument(
+        "--uninstall-auto-updates",
+        dest="uninstall_auto_updates",
         action="store_true",
-        help="Disable adaptive rate limiting",
+        help="Uninstall all Homebrew casks that have auto-updates",
     )
 
-    # Filtering options
-    filter_group = parser.add_argument_group("Filtering options")
-    # Deprecated: --blacklist retained for backward compatibility (use --blocklist)
-    filter_group.add_argument(
-        "--blacklist",
-        dest="blacklist",
-        help="(DEPRECATED) Comma-separated list of applications to ignore (use --blocklist instead)",
-    )
-    filter_group.add_argument(
-        "--blocklist",
-        dest="blocklist",
-        help="Comma-separated list of applications to ignore",
-    )
-    filter_group.add_argument(
-        "--additional-dirs",
-        dest="additional_dirs",
-        help="Colon-separated list of additional directories to scan for applications",
-    )
-    filter_group.add_argument(
-        "--similarity",
-        dest="similarity",
-        type=int,
-        help="Similarity threshold for matching (0-100, default: 75)",
-    )
-    filter_group.add_argument(
-        "--no-enhanced-matching",
-        dest="no_enhanced_matching",
-        action="store_true",
-        help="Disable enhanced fuzzy matching (use basic matching instead)",
-    )
-    filter_group.add_argument(
+    # Filter options for auto-updates
+    parser.add_argument(
         "--exclude-auto-updates",
         dest="exclude_auto_updates",
         action="store_true",
-        help="Exclude applications that have auto-updates enabled in Homebrew",
+        help="Exclude applications that have auto-updates",
     )
-    filter_group.add_argument(
+    parser.add_argument(
         "--only-auto-updates",
         dest="only_auto_updates",
         action="store_true",
-        help="Only show applications that have auto-updates enabled in Homebrew",
+        help="Only show applications that have auto-updates",
     )
 
-    # Filter management
-    filter_management_group = parser.add_argument_group("Filter management")
-    filter_management_group.add_argument(
-        "--save-filter",
-        dest="save_filter",
-        metavar="NAME",
-        help="Save current filter settings with the given name",
-    )
-    filter_management_group.add_argument(
-        "--load-filter",
-        dest="load_filter",
-        metavar="NAME",
-        help="Load saved filter settings with the given name",
-    )
-    filter_management_group.add_argument(
-        "--list-filters",
-        dest="list_filters",
+    # Service management options
+    parser.add_argument(
+        "--install-service",
+        dest="install_service",
         action="store_true",
-        help="List all saved filters",
+        help="Install macOS service",
     )
-    filter_management_group.add_argument(
-        "--delete-filter",
-        dest="delete_filter",
-        metavar="NAME",
-        help="Delete a saved filter",
+    parser.add_argument(
+        "--uninstall-service",
+        dest="uninstall_service",
+        action="store_true",
+        help="Uninstall macOS service",
     )
-
-    # Export options
-    export_group = parser.add_argument_group("Export options")
-    export_group.add_argument(
-        "--export",
-        dest="export_format",
-        choices=["json", "csv"],
-        help="Export results in specified format (json or csv)",
+    parser.add_argument(
+        "--service-status",
+        dest="service_status",
+        action="store_true",
+        help="Check macOS service status",
     )
-    export_group.add_argument(
-        "--output-file",
-        dest="output_file",
-        help="Specify the output file for export (default: print to stdout)",
+    parser.add_argument(
+        "--test-notification",
+        dest="test_notification",
+        action="store_true",
+        help="Test macOS notification",
+    )
+    parser.add_argument(
+        "--menubar",
+        action="store_true",
+        help="Launch menubar app",
     )
 
     # Configuration options
-    config_group = parser.add_argument_group("Configuration options")
-    config_group.add_argument(
+    parser.add_argument(
+        "--config",
+        "-c",
+        type=str,
+        help="Path to configuration file",
+    )
+    parser.add_argument(
         "--generate-config",
         dest="generate_config",
         action="store_true",
-        help="Generate a default configuration file at ~/.config/versiontracker/config.yaml",
+        help="Generate default configuration file",
     )
-    config_group.add_argument(
-        "--config-path",
-        dest="config_path",
-        help=(
-            "Specify an alternative path for the configuration file (can be used both for "
-            "generating a config file with --generate-config and for using a custom config "
-            "file location when running the application)"
-        ),
+
+    # Export options
+    parser.add_argument(
+        "--export",
+        dest="export_format",
+        choices=["json", "yaml", "csv"],
+        help="Export results in specified format",
     )
-    config_group.add_argument(
-        "--service-interval",
-        dest="service_interval",
+
+    # Filter management
+    parser.add_argument(
+        "--save-filter",
+        dest="save_filter",
+        type=str,
+        help="Save current query as named filter",
+    )
+
+    # Debugging and profiling
+    parser.add_argument(
+        "--debug",
+        "-d",
+        action="count",
+        default=0,
+        help="Enable debug output (use -dd for verbose debug)",
+    )
+    parser.add_argument(
+        "--profile",
+        action="store_true",
+        help="Enable performance profiling",
+    )
+    parser.add_argument(
+        "--detailed-profile",
+        dest="detailed_profile",
+        action="store_true",
+        help="Show detailed profiling information",
+    )
+
+    # Other common options
+    parser.add_argument(
+        "--verbose",
+        "-v",
+        action="store_true",
+        help="Enable verbose output",
+    )
+    parser.add_argument(
+        "--max-workers",
+        dest="max_workers",
         type=int,
-        default=24,
-        help="Interval in hours for scheduled service (default: 24)",
+        help="Maximum number of worker threads",
     )
-    config_group.add_argument(
-        "--notify",
-        dest="notify",
+    parser.add_argument(
+        "--rate-limit",
+        dest="rate_limit",
+        type=float,
+        help="Rate limit for network requests",
+    )
+    parser.add_argument(
+        "--blacklist",
+        type=str,
+        help="Blacklist of applications to exclude",
+    )
+    parser.add_argument(
+        "--blocklist",
+        type=str,
+        help="Blocklist of applications to exclude",
+    )
+    parser.add_argument(
+        "--brew-filter",
+        dest="brew_filter",
         action="store_true",
-        help="Send macOS notification with results",
+        help="Filter out Homebrew casks from results",
+    )
+    parser.add_argument(
+        "--no-progress",
+        dest="no_progress",
+        action="store_true",
+        help="Disable progress bar",
+    )
+    parser.add_argument(
+        "--similarity",
+        type=float,
+        help="Similarity threshold for matching",
+    )
+    parser.add_argument(
+        "--additional-dirs",
+        dest="additional_dirs",
+        type=str,
+        help="Additional directories to search",
     )
 
-    # Main command group (mutually exclusive)
-    group = parser.add_mutually_exclusive_group()
-    group.add_argument(
-        "-a",
-        "--apps",
-        action="store_true",
-        dest="apps",
-        help="return Apps in Applications/ that is not updated by App Store",
-    )
-    group.add_argument(
-        "-b",
-        "--brews",
-        action="store_true",
-        dest="brews",
-        help="return installable brews",
-    )
-    group.add_argument(
-        "-r",
-        "--recommend",
-        action="store_true",
-        dest="recom",
-        help="return recommendations for brew",
-    )
-    group.add_argument(
-        "--strict-recommend",
-        action="store_true",
-        dest="strict_recom",
-        help=("return recommendations for brew, excluding apps that can already be installed with brew"),
-    )
-    group.add_argument(
-        "-o",
-        "--outdated",
-        action="store_true",
-        dest="check_outdated",
-        help="check for outdated applications compared to Homebrew versions",
-    )
-    # Deprecated: --blacklist-auto-updates retained for backward compatibility
-    group.add_argument(
-        "--blacklist-auto-updates",
-        action="store_true",
-        dest="blacklist_auto_updates",
-        help="(DEPRECATED) Add all applications with auto-updates to the blocklist (use --blocklist-auto-updates)",
-    )
-    group.add_argument(
-        "--blocklist-auto-updates",
-        action="store_true",
-        dest="blocklist_auto_updates",
-        help="Add all applications with auto-updates to the blocklist",
-    )
-    group.add_argument(
-        "--uninstall-auto-updates",
-        action="store_true",
-        dest="uninstall_auto_updates",
-        help="Uninstall all Homebrew casks that have auto-updates enabled (with confirmation)",
-    )
-    group.add_argument(
-        "--install-service",
-        action="store_true",
-        dest="install_service",
-        help="Install macOS scheduled checker service (launchd)",
-    )
-    group.add_argument(
-        "--uninstall-service",
-        action="store_true",
-        dest="uninstall_service",
-        help="Uninstall macOS scheduled checker service",
-    )
-    group.add_argument(
-        "--service-status",
-        action="store_true",
-        dest="service_status",
-        help="Show status of macOS scheduled checker service",
-    )
-    group.add_argument(
-        "--test-notification",
-        action="store_true",
-        dest="test_notification",
-        help="Send a test macOS notification",
-    )
-    group.add_argument(
-        "--menubar",
-        action="store_true",
-        dest="menubar",
-        help="Launch macOS menubar application",
-    )
-    group.add_argument("-V", "--version", action="version", version=f"%(prog)s {__version__}")
+    return parser.parse_args()
 
-    # If no arguments were provided, show help
-    if len(sys.argv) == 1:
-        parser.print_help()
-        sys.exit(1)
 
-    args = parser.parse_args()
-
-    # Deprecation handling / value mapping (centralized via deprecation utility)
-    if getattr(args, "blacklist", None) and not getattr(args, "blocklist", None):
-        args.blocklist = args.blacklist
-        warn_deprecated_flag(
-            flag="--blacklist",
-            replacement="--blocklist",
-            removal_version="0.8.0",
-        )
-
-    if getattr(args, "blacklist_auto_updates", None) and not getattr(args, "blocklist_auto_updates", None):
-        args.blocklist_auto_updates = args.blacklist_auto_updates
-        warn_deprecated_flag(
-            flag="--blacklist-auto-updates",
-            replacement="--blocklist-auto-updates",
-            removal_version="0.8.0",
-        )
-
-    return args
+if __name__ == "__main__":
+    # Provide default arguments for main() if run directly
+    main(None, False)
